// src/app/editor/enums/property-name.ts
export enum PropertyName {
    ControlType = 'ControlType',
    Name = 'Name',
    Caption = 'Caption',
    Left = 'Left',
    Top = 'Top',
    Height = 'Height',
    Width = 'Width',
    Bold = 'Bold',
    CarePlans = 'Care Plans',
    Children = 'Children',
    ControlGroup = 'Control Group',
    ClearAll = 'Clear All',
    DisableAll = 'Disable All',
    OASISName = 'OASIS Name',
    OASISLabel = 'OASIS Label',
    OASISCode = 'OASIS Code',
    Rules = 'Rules',
    Parent = 'Parent'

}